+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+- Added Dazzling Display Animation and Automation
+- Tectonic Stomp Automation and Animation
+
 ## [0.56.0] - 2025-09-22
 
 - Frightened & Antagonized Automations